Move ICMP extension helpers into their own file
