+++ src/icmp.cpp
@@ -36,8 +36,8 @@
 #include "utils.h"
 #include "exceptions.h"
 #include "icmp.h"
-#include "internals.h"
 #include "memory_helpers.h"
+#include "detail/icmp_extension_helpers.h"
 
 using std::memset;
 using std::max;
